requirement.txt
torch
torchvision
onnxruntime-gpu==1.12.0
onnxruntime==1.12.0
numpy==1.23.5
collections
scipy==1.10.0
torchlibrosa==0.1.0
python==3.8.10
